+++ exorad/models/instruments/__init__.py
@@ -1,3 +1,3 @@
+from .instrument import Instrument
 from .photometer import Photometer
 from .spectrometer import Spectrometer
-from .instrument import Instrument